Add getting users' json by id

+++ src/main/java/pl/bartekde/booktradingspring/dao/UserDao.java
@@ -5,8 +5,8 @@ import pl.bartekde.booktradingspring.entity.User;
 public interface UserDao {
 
     User findByUsername(String username);
-    
-    User findById(int id);
+
+    User findById(long id);
 
     void save(User user);
 }
